more IMU adjustment, get yaw from gyro, more ESC testing code

+++ stm32_f103/Core/Src/stm32f1xx_it.c
@@ -53,7 +53,7 @@
 
 /* Private user code ---------------------------------------------------------*/
 /* USER CODE BEGIN 0 */
-extern int16_t position;
+extern uint32_t position;
 int16_t oldpos = 0;
 extern double speed;
 int indx = 0;
